+++ packages/umi-build-dev/src/getPlugins.js
@@ -23,8 +23,8 @@ export default function(opts = {}) {
     './plugins/commands/generate',
     './plugins/commands/rm',
     './plugins/commands/config',
-    './plugins/commands/block',
     ...(process.env.UMI_UI === 'none' ? [] : [require.resolve('umi-plugin-ui')]),
+    './plugins/commands/block',
     './plugins/commands/version',
     './plugins/global-js',
     './plugins/global-css',
