+++ packages/umi-build-dev/src/plugins/commands/block/ui/flagBabelPlugin/fixtures/body-jsxelement/expected.js
