+++ packages/umi-build-dev/src/plugins/commands/block/ui/flagBabelPlugin/fixtures/export-function/origin.js
