+++ packages/umi-build-dev/src/plugins/commands/block/ui/locales/en-US.js
