+++ packages/umi-build-dev/src/plugins/commands/block/ui/locales/zh-CN.js
