+++ packages/umi-plugin-ui/src/index.ts
@@ -21,5 +21,4 @@ require('${join(__dirname, '../bubble')}').default({
   require('./plugins/dashboard/index').default(api);
   require('./plugins/configuration/index').default(api);
   require('./plugins/tasks/index').default(api);
-  require('./plugins/blocks/index').default(api);
 };
